Refactor for dedicated audio playback component, refactor components to use const fns
